+++ app/components/templates/panel/dashboard-entity-panel.actions.tsx
@@ -9,7 +9,7 @@ import {
 	type IDashboardPanelEntityActionStrategy,
 	type IPanelEntityActionStrategy,
 } from '#app/strategies/component/dashboard-panel/entity-action/entity-action'
-import { type IDashboardPanelSelectEntityStrategy } from '#app/strategies/component/dashboard-panel/select-entity.strategy'
+import { type IDashboardPanelSelectEntityStrategy } from '#app/strategies/component/dashboard-panel/update-entity-selected.strategy'
 import { type IDashboardPanelUpdateEntityVisibleStrategy } from '#app/strategies/component/dashboard-panel/update-entity-visible.strategy'
 import { SidebarPanelRowActionsContainer } from '..'
 import { PanelEntityDeleteAction } from './dashboard-entity-panel.actions.delete'
