+++ app/strategies/component/dashboard-panel/update-entity-selected.strategy.ts
